feat: 컴포넌트, 기능으로 구조화
