+++ src/features/cart/index.js
@@ -1,0 +1,2 @@
+export * from "./cart.state.js";
+export { openCartModal } from "./cart.modal.js";
