+++ src/pages/HomePage.js
@@ -1,45 +1,5 @@
-const LoadingCard = `
-<div class="bg-white rounded-lg shadow-sm border border-gray-200 overflow-hidden animate-pulse">
-<div class="aspect-square bg-gray-200"></div>
-<div class="p-3">
-    <div class="h-4 bg-gray-200 rounded mb-2"></div>
-    <div class="h-3 bg-gray-200 rounded w-2/3 mb-2"></div>
-    <div class="h-5 bg-gray-200 rounded w-1/2 mb-3"></div>
-    <div class="h-8 bg-gray-200 rounded"></div>
-</div>
-</div>`;
-
-const LoadingCardList = LoadingCard.repeat(4);
-
-const ProductCard = (product) => `
- <div class="bg-white rounded-lg shadow-sm border border-gray-200 overflow-hidden product-card"
-    data-product-id="${product.productId}">
-<!-- 상품 이미지 -->
-<div class="aspect-square bg-gray-100 overflow-hidden cursor-pointer product-image">
-    <img src="${product.image}"
-        alt="${product.title}"
-        class="w-full h-full object-cover hover:scale-105 transition-transform duration-200"
-        loading="lazy">
-</div>
-<!-- 상품 정보 -->
-<div class="p-3">
-    <div class="cursor-pointer product-info mb-3">
-    <h3 class="text-sm font-medium text-gray-900 line-clamp-2 mb-1">
-        ${product.title}
-    </h3>
-    <p class="text-xs text-gray-500 mb-2"></p>
-    <p class="text-lg font-bold text-gray-900">
-        ${product.lprice}원
-    </p>
-    </div>
-    <!-- 장바구니 버튼 -->
-    <button class="w-full bg-blue-600 text-white text-sm py-2 px-3 rounded-md
-            hover:bg-blue-700 transition-colors add-to-cart-btn" data-product-id="${product.productId}">
-    장바구니 담기
-    </button>
-</div>
-</div>
-`;
+import ProductCard from "../components/ProductCard.js";
+import { LoadingCardList } from "../components/LoadingCard.js";
 
 export default function HomePage({ products = [], total = 0, loading = false, loadingMore = false, categories = [] }) {
   return 상품목록_레이아웃_로딩(products, total, loading, loadingMore, categories);
@@ -91,7 +51,12 @@ const 상품목록_레이아웃_로딩 = (products, total, loading, loadingMore)
               </div>
               <!-- 1depth 카테고리 -->
               <div class="flex flex-wrap gap-2">
-              ${loading ? `<div class="text-sm text-gray-500 italic">카테고리 로딩 중...</div>` : ""}
+              ${
+                loading
+                  ? `<div class="text-sm text-gray
+500 italic">카테고리 로딩 중...</div>`
+                  : ""
+              }
               </div>
               <!-- 2depth 카테고리 -->
             </div>
